requirements.txt
asgiref==3.7.2
Django==4.2.11
django-cors-headers==4.3.1
ffmpeg-python==0.2.0
future==1.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
